[ZEPPELIN-5031]. Introduce zeppelin-common module
### What is this PR for?

This is a refactoring PR which introduce zeppelin-common module that contains the common code for both zeppelin server and client api.

### What type of PR is it?
[Refactoring]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-5031

### How should this be tested?
* First time? Setup Travis CI as described on https://zeppelin.apache.org/contribution/contributions.html#continuous-integration
* Strongly recommended: add automated unit tests for any new or changed behavior
* Outline any manual steps to test the PR here.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update?
* Is there breaking changes for older versions?
* Does this needs documentation?

Author: Jeff Zhang <zjffdu@apache.org>

Closes #3900 from zjffdu/ZEPPELIN-5031 and squashes the following commits:

65865d350 [Jeff Zhang] [ZEPPELIN-5031]. Introduce zeppelin-common module

(cherry picked from commit 847e12db3e61b19fd0301281df17424b8572601f)
