+++ pom.xml
@@ -93,6 +93,7 @@
     <module>geode</module>
     <module>ksql</module>
     <module>sparql</module>
+    <module>zeppelin-common</module>
     <module>zeppelin-client</module>
     <module>zeppelin-client-examples</module>
     <module>zeppelin-web</module>
