+++ zeppelin-client/src/main/java/org/apache/zeppelin/client/websocket/AbstractMessageHandler.java
@@ -17,6 +17,8 @@
 
 package org.apache.zeppelin.client.websocket;
 
+import org.apache.zeppelin.common.Message;
+
 import com.google.gson.Gson;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
