+++ zeppelin-client/src/main/java/org/apache/zeppelin/client/websocket/ZeppelinWebSocketClient.java
@@ -18,6 +18,8 @@
 package org.apache.zeppelin.client.websocket;
 
 import com.google.gson.Gson;
+import org.apache.zeppelin.common.Message;
+
 import org.eclipse.jetty.websocket.api.Session;
 import org.eclipse.jetty.websocket.api.annotations.OnWebSocketClose;
 import org.eclipse.jetty.websocket.api.annotations.OnWebSocketConnect;
