+++ zeppelin-common/src/main/java/org/apache/zeppelin/common/JsonSerializable.java
